Make PDFRenderer properly RESTful
The PDFRenderer has, until now, been running as an additional
detail view in views, with a URL change needed to call it.

This update moves the code from that detail view into the
renderer where it belongs, meaning the PDF comes from the
same endpoint by varying the 'Content-Type' header in the
client.